relocated all style files from components folder to styles folder
