+++ client/src/components/image-slider.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import "./image-slider.css";
+import "../styles/components/image-slider.css";
 import testbild1 from '../images/testbildx.jpg';
 import testbild2 from '../images/testfeatures.jpg';
 import testbild3 from '../images/testbildy.jpg';
@@ -20,13 +20,13 @@ function ImageComponent() {
           <div id="slide-1"><img src={testbild1} style={{width:"100%"}}/></div>
           <div id="slide-2"><img src={testbild2} style={{width:"100%"}}/></div>
           <div id="slide-3"><img src={testbild3} style={{width:"100%"}}/></div>
-        
+
         </div>
         <a href="#slide-1"><img src={testbild1} style={{width:"100%"}}/></a>
         <a href="#slide-2"><img src={testbild2} style={{width:"100%"}}/></a>
         <a href="#slide-3"><img src={testbild3} style={{width:"100%"}}/></a>
       </div>
-   
+
     </div>
   )
 }
