+++ client/src/components/navbar.js
@@ -3,7 +3,7 @@ import { useState } from "react"
 import { Link, useOutlet } from "react-router-dom"
 import { useLogout } from "../hooks/useLogout"
 import { useAuthContext } from "../hooks/useAuthContext"
-import styles from "./navbar.module.scss"
+import styles from "../styles/components/navbar.module.scss"
 import avatars from "../styles/avatars.module.scss"
 import buttons from "../styles/buttons.module.scss"
 import BurgerButton from "./atoms/burgerButton"
