+++ client/src/styles/components/footer.module.scss
@@ -1,5 +1,5 @@
-@import '../styles/colors.module.scss';
-@import '../styles/fonts.module.scss';
+@import '../colors.module.scss';
+@import '../fonts.module.scss';
 
 a {
   &:active,
