+++ client/src/styles/components/image-slider.css
