+++ client/src/styles/components/intro-block.module.scss
@@ -1,5 +1,5 @@
-@import '../styles/colors.module.scss';
-@import '../styles/fonts.module.scss';
+@import '../colors.module.scss';
+@import '../fonts.module.scss';
 
 * {
   box-sizing: border-box;
